requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.4
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
einops==0.6.1
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
galore-torch==1.0
gdown==4.6.0
h5py==3.13.0
humanfriendly==10.0
icecream==2.1.3
imagecodecs @ file:///croot/imagecodecs_1734436728732/work
imageio @ file:///croot/imageio_1738159938990/work
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader @ file:///croot/lazy_loader_1718176737906/work
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.9.4
MedPy==0.4.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_collections==0.1.1
monai==1.1.0
mpmath==1.3.0
networkx @ file:///croot/networkx_1717597493534/work
nltk==3.9.1
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.5.4.58
opencv-python-headless==4.10.0
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp39-cp39-linux_x86_64.whl#sha256=0a51ed2e81ab863e3d00ed6c5049192ce578ecb38fb467d2f9a6585d3c25f666
pillow @ file:///croot/pillow_1738010226202/work
ply==3.11
protobuf==3.20.3
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5_sip @ file:///croot/pyqt-split_1736540531116/work/pyqt_sip
PySocks==1.7.1
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
rouge_score==0.1.2
safetensors==0.5.3
scikit-image @ file:///croot/scikit-image_1726737398091/work
scikit-learn==1.6.1
scipy @ file:///croot/scipy_1733756309941/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=a4372e07d982a377bda57ec71fd74ac403442571e1d717de47325a14dea5c278
seaborn==0.13.2
SimpleITK==2.2.1
sip @ file:///croot/sip_1736541134699/work
six @ file:///croot/six_1744271502820/work
soupsieve==2.6
sympy==1.13.1
tensorboardX==2.6
tensorly==0.9.0
threadpoolctl==3.6.0
tifffile @ file:///croot/tifffile_1695107451082/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado @ file:///croot/tornado_1733960490606/work
tqdm @ file:///croot/tqdm_1738943501192/work
triton==3.2.0
typing_extensions==4.13.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
zipp @ file:///croot/zipp_1732630741423/work
